removes redundant commitment computation, plus clippy fixes

+++ spartan/src/r1csinstance.rs
@@ -49,8 +49,8 @@ impl<G: CurveGroup> R1CSCommitmentGens<G> {
     num_nz_entries: usize,
   ) -> R1CSCommitmentGens<G> {
     assert!(num_inputs < num_vars);
-    let num_poly_vars_x = num_cons.log_2() as usize;
-    let num_poly_vars_y = (2 * num_vars).log_2() as usize;
+    let num_poly_vars_x = num_cons.log_2();
+    let num_poly_vars_y = (2 * num_vars).log_2();
     let gens =
       SparseMatPolyCommitmentGens::new(label, num_poly_vars_x, num_poly_vars_y, num_nz_entries, 3);
     R1CSCommitmentGens { gens }
@@ -118,8 +118,8 @@ impl<F: PrimeField> R1CSInstance<F> {
     assert!(num_inputs < num_vars);
 
     // no errors, so create polynomials
-    let num_poly_vars_x = num_cons.log_2() as usize;
-    let num_poly_vars_y = (2 * num_vars).log_2() as usize;
+    let num_poly_vars_x = num_cons.log_2();
+    let num_poly_vars_y = (2 * num_vars).log_2();
 
     let mat_A = (0..A.len())
       .map(|i| SparseMatEntry::new(A[i].0, A[i].1, A[i].2))
@@ -169,8 +169,8 @@ impl<F: PrimeField> R1CSInstance<F> {
     let mut prng = test_rng();
 
     // assert num_cons and num_vars are power of 2
-    assert_eq!((num_cons.log_2() as usize).pow2(), num_cons);
-    assert_eq!((num_vars.log_2() as usize).pow2(), num_vars);
+    assert_eq!((num_cons.log_2()).pow2(), num_cons);
+    assert_eq!((num_vars.log_2()).pow2(), num_vars);
 
     // num_inputs + 1 <= num_vars
     assert!(num_inputs < num_vars);
@@ -215,8 +215,8 @@ impl<F: PrimeField> R1CSInstance<F> {
     Timer::print(&format!("number_non-zero_entries_B {}", B.len()));
     Timer::print(&format!("number_non-zero_entries_C {}", C.len()));
 
-    let num_poly_vars_x = num_cons.log_2() as usize;
-    let num_poly_vars_y = (2 * num_vars).log_2() as usize;
+    let num_poly_vars_x = num_cons.log_2();
+    let num_poly_vars_y = (2 * num_vars).log_2();
     let poly_A = SparseMatPolynomial::new(num_poly_vars_x, num_poly_vars_y, A);
     let poly_B = SparseMatPolynomial::new(num_poly_vars_x, num_poly_vars_y, B);
     let poly_C = SparseMatPolynomial::new(num_poly_vars_x, num_poly_vars_y, C);
